+++ changelog.md
@@ -1,4 +1,13 @@
 # Changelog
+### 2020-03-03
+#### enable GUIs for events
+  * new method of eventPresenter isEnabledForEvent($event):bool
+    * enable GUIs for single events
+  * Payments, Applications (single+team), applications, fyziklani,...
+  * Remove Application progress presenter
+    * progress charts are moved to list view of application presenter
+  
+----
 ### 2020-02-29
 #### Email_messages
   * add email_message detail GUI